feat: textmgmt more necessary restrictions on the api implemented

+++ TEQST/textmgmt/models.py
@@ -5,6 +5,11 @@ from .utils import folder_path, folder_relative_path
 import os
 import shutil
 
+################################
+# important todos:
+# - before text ceration a sharedfolder must have been created
+################################
+
 
 class Folder(models.Model):
     name = models.CharField(max_length=50)
@@ -31,10 +36,15 @@ class Folder(models.Model):
             if f_old.name != self.name:  # if name changed
                 os.rename(path_f_old, path)
         super().save(*args, **kwargs)
+        if self.is_shared_folder() and not isinstance(self, SharedFolder):
+            sf = self.sharedfolder
+            sf.name = self.name
+            sf.save()
     
     def delete(self, *args, **kwargs):
         # TODO maybe this is desired, maybe not.
-        shutil.rmtree(folder_path(self))
+        if not isinstance(self, SharedFolder):
+            shutil.rmtree(folder_path(self))
         # does not work if the folder is not there
         super().delete(*args, **kwargs)
     
@@ -44,22 +54,23 @@ class Folder(models.Model):
     def make_shared_folder(self):
         #Check for folder having no children is required:
         # len(self.subfolder.all()) == 0
-        if self.is_shared_folder():
+        if isinstance(self, SharedFolder):
             return self
+        if self.is_shared_folder():
+            return self.sharedfolder
         sf = SharedFolder(folder_ptr=self, name=self.name, owner=self.owner, parent=self.parent)
         sf.save()
         return sf
 
 
 class SharedFolder(Folder):
-    # "null has no effect"
-    speaker = models.ManyToManyField(settings.AUTH_USER_MODEL, related_name='sharedfolder', null=True, blank=True)
+    speaker = models.ManyToManyField(settings.AUTH_USER_MODEL, related_name='sharedfolder', blank=True)
 
     # __str__() is inherited
 
     def save(self, *args, **kwargs):
         # actual folder is created in super method
-        # TODO create other necessary stuff like temp folders etc.
+        # TODO upon creation create other necessary stuff like temp folders etc.
         super().save(*args, **kwargs)
 
     def get_path(self):
@@ -79,20 +90,15 @@ class Text(models.Model):
     # sentences_count = models.IntegerField(null=True)
     # sentences = models.TextField(null=True)
     # sentences can be replaced. A file read can easily be done from within the model
-    shared_folder = models.ForeignKey(SharedFolder, on_delete=models.CASCADE, related_name='text')
+    shared_folder = models.ForeignKey(Folder, on_delete=models.CASCADE, related_name='text')
     textfile = models.FileField(upload_to=upload_path)
 
     def __str__(self):
         return self.title
     
     def save(self, *args, **kwargs):
-        # TODO open file, fill fields sentences, sentences_count, close file
-        # if not self.pk:
-        #     count = 1
-        #     self.textfile.open('r')
-        #     for line in self.textfile:
-        #         count += 1
-        #     self.sentences_count = int(count / 2)
+        # TODO a sharedfolder must have been created
+        self.shared_folder = self.shared_folder.make_shared_folder()
         super().save(*args, **kwargs)
     
     def get_content(self):
